updated context for CustomerContext
