+++ ecommerce-client/src/components/Customers/Customers.tsx
@@ -1,14 +1,14 @@
-import { useEffect, useReducer, useState } from "react"
-import { CustomerContext } from "../../contexts/customerContext";
+import { useContext, useEffect, useState } from "react"
 import { useCustomers } from "../../hooks/useCustomers";
-import { CustomerReducer, ActionType } from "../../reducers/CustomerReducer";
+import { ActionType } from "../../reducers/CustomerReducer";
 import { CreateCustomer } from "./CreateCustomer";
 import { UpdateCustomer } from "./UpdateCustomer";
+import CustomerContext from "../../contexts/CustomerContext";
 
 
 export const Customers = () => {
     const { fetchCustomersHandler, deleteCustomerHandler } = useCustomers()
-    const [customers, dispatch] = useReducer(CustomerReducer, []);
+    const { customers, dispatch } = useContext(CustomerContext)
     const [editingCustomerId, setEditingCustomerId] = useState<number | null>(null);
 
     useEffect(() => {
@@ -32,7 +32,7 @@ export const Customers = () => {
     }
 
     return (
-        <CustomerContext.Provider value={{ customers, dispatch }}>
+        <>
             <CreateCustomer />
             <section>
                 {customers.map(({ id, firstname, lastname, city }) => (
@@ -49,14 +49,10 @@ export const Customers = () => {
                                     <button onClick={() => handleDelete(id)}>Delete</button>
                                 </div>
                             </>
-                        )
-
-
-                        }
+                        )}
                     </div>
-                )
-                )}
+                ))}
             </section>
-        </CustomerContext.Provider>
+        </>
     )
 }
