+++ ecommerce-client/src/components/Customers/UpdateCustomer.tsx
@@ -1,9 +1,9 @@
 import { ChangeEvent, FormEvent, useContext, useEffect, useState } from "react"
-import { CustomerContext } from "../../contexts/customerContext";
 import { useCustomers } from "../../hooks/useCustomers";
 import { ActionType } from "../../reducers/CustomerReducer";
 import { updateCustomer } from "../../services/customerService";
 import { Customer } from "../../types/Customer";
+import CustomerContext from "../../contexts/CustomerContext";
